[WIP] finish adapter refactor
