+++ src/devTools/editor/extensionPoints/menuItem.ts
@@ -15,13 +15,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { IExtension, Metadata } from "@/core";
+import { IconConfig, IExtension, Metadata } from "@/core";
 import {
   ButtonDefinition,
   ButtonSelectionResult,
 } from "@/nativeEditor/insertButton";
 import { FrameworkMeta } from "@/messaging/constants";
-import { ActionFormState } from "@/devTools/editor/editorSlice";
 import {
   makeBaseState,
   makeExtensionReaders,
@@ -34,6 +33,7 @@ import {
 import {
   MenuDefinition,
   MenuItemExtensionConfig,
+  MenuPosition,
 } from "@/extensionPoints/menuItemExtension";
 import { ExtensionPointConfig } from "@/extensionPoints/types";
 import { castArray, identity, pickBy } from "lodash";
@@ -49,7 +49,12 @@ import { v4 as uuidv4 } from "uuid";
 import { getDomain } from "@/permissions/patterns";
 import { faMousePointer } from "@fortawesome/free-solid-svg-icons";
 import * as nativeOperations from "@/background/devtools";
-import { ElementConfig } from "@/devTools/editor/extensionPoints/elementConfig";
+import {
+  BaseFormState,
+  ElementConfig,
+} from "@/devTools/editor/extensionPoints/elementConfig";
+import { ElementInfo } from "@/nativeEditor/frameworks";
+import { BlockPipeline } from "@/blocks/combinators";
 
 export const wizard: WizardStep[] = [
   { step: "Name", Component: MetaTab },
@@ -62,7 +67,38 @@ export const wizard: WizardStep[] = [
   { step: "Logs", Component: LogsTab },
 ];
 
-export function makeActionState(
+export interface ActionFormState extends BaseFormState {
+  type: "menuItem";
+
+  containerInfo: ElementInfo;
+
+  extensionPoint: {
+    metadata: Metadata;
+    definition: {
+      containerSelector: string;
+      position?: MenuPosition;
+      template: string;
+      isAvailable: {
+        matchPatterns: string;
+        selectors: string;
+      };
+    };
+    traits?: {
+      style: {
+        mode: "default" | "inherit";
+      };
+    };
+  };
+
+  extension: {
+    caption: string;
+    dynamicCaption?: boolean;
+    icon?: IconConfig;
+    action: BlockPipeline;
+  };
+}
+
+function initialFormStateFactory(
   url: string,
   metadata: Metadata,
   button: ButtonSelectionResult,
@@ -98,7 +134,7 @@ export function makeActionState(
   };
 }
 
-export function makeMenuExtensionPoint({
+function selectExtensionPoint({
   extensionPoint,
   readers,
 }: ActionFormState): ExtensionPointConfig<MenuDefinition> {
@@ -127,7 +163,7 @@ export function makeMenuExtensionPoint({
   };
 }
 
-export function makeActionExtension({
+function selectExtension({
   uuid,
   label,
   extensionPoint,
@@ -182,7 +218,7 @@ export async function makeActionExtensionFormState(
   };
 }
 
-export async function makeActionFormState(
+export async function fromExtension(
   config: IExtension<MenuItemExtensionConfig>
 ): Promise<ActionFormState> {
   const extensionPoint = await lookupExtensionPoint<
@@ -220,8 +256,8 @@ export async function makeActionFormState(
 function asDynamicElement(element: ActionFormState): ButtonDefinition {
   return {
     type: "menuItem",
-    extension: makeActionExtension(element),
-    extensionPoint: makeMenuExtensionPoint(element),
+    extension: selectExtension(element),
+    extensionPoint: selectExtensionPoint(element),
     readers: makeExtensionReaders(element),
   };
 }
@@ -231,11 +267,11 @@ const config: ElementConfig<ButtonSelectionResult, ActionFormState> = {
   label: "Button",
   icon: faMousePointer,
   insert: nativeOperations.insertButton,
-  makeState: makeActionState,
+  initialFormStateFactory,
   asDynamicElement,
-  extensionPoint: makeMenuExtensionPoint,
-  extension: makeActionExtension,
-  formState: makeActionFormState,
+  selectExtensionPoint,
+  selectExtension,
+  fromExtension,
 };
 
 export default config;
