+++ src/devTools/editor/hooks/useCreate.ts
@@ -15,11 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {
-  FormState,
-  editorSlice,
-  isCustomReader,
-} from "@/devTools/editor/editorSlice";
+import { FormState, editorSlice } from "@/devTools/editor/editorSlice";
 import { useDispatch } from "react-redux";
 import { useCallback, useState } from "react";
 import axios, { AxiosError, AxiosRequestConfig } from "axios";
@@ -40,6 +36,7 @@ import { reportEvent } from "@/telemetry/events";
 import { removeUndefined } from "@/utils";
 import { fromJS as extensionPointFactory } from "@/extensionPoints/factory";
 import { ensureAllPermissions, extensionPermissions } from "@/permissions";
+import { isCustomReader } from "@/devTools/editor/extensionPoints/elementConfig";
 
 const { saveExtension } = optionsSlice.actions;
 const { markSaved } = editorSlice.actions;
@@ -222,7 +219,7 @@ export function useCreate(): CreateCallback {
           )
         ) {
           try {
-            const extensionPointConfig = adapter.extensionPoint(element);
+            const extensionPointConfig = adapter.selectExtensionPoint(element);
             const packageId = element.installed
               ? editablePackages.find(
                   (x) => x.name === extensionPointConfig.metadata.id
@@ -249,7 +246,7 @@ export function useCreate(): CreateCallback {
         }
 
         try {
-          dispatch(saveExtension(adapter.extension(element)));
+          dispatch(saveExtension(adapter.selectExtension(element)));
           dispatch(markSaved(element.uuid));
           void reactivate();
           addToast("Saved extension", {
